+++ core-ng/src/main/java/core/framework/impl/db/DBEnumMapper.java
@@ -1,6 +1,6 @@
 package core.framework.impl.db;
 
-import core.framework.api.db.EnumValue;
+import core.framework.api.db.DBEnumValue;
 import core.framework.api.util.Exceptions;
 import core.framework.api.util.Maps;
 
@@ -34,7 +34,7 @@ final class DBEnumMapper<T extends Enum<T>> {
         for (T constant : constants) {
             try {
                 Field field = enumClass.getField(constant.name());
-                String dbValue = field.getDeclaredAnnotation(EnumValue.class).value();
+                String dbValue = field.getDeclaredAnnotation(DBEnumValue.class).value();
                 mapping.put(dbValue, constant);
             } catch (NoSuchFieldException e) {
                 throw new Error(e);