+++ core-ng/src/main/java/core/framework/impl/db/DatabaseClassValidator.java
@@ -1,7 +1,7 @@
 package core.framework.impl.db;
 
 import core.framework.api.db.Column;
-import core.framework.api.db.EnumValue;
+import core.framework.api.db.DBEnumValue;
 import core.framework.api.db.PrimaryKey;
 import core.framework.api.db.Table;
 import core.framework.api.util.Exceptions;
@@ -98,13 +98,21 @@ final class DatabaseClassValidator implements TypeVisitor {
 
     private void validateEnumClass(Class<?> enumClass, Field field) {
         Enum[] constants = (Enum[]) enumClass.getEnumConstants();
+        Set<String> enumValues = Sets.newHashSet();
         for (Enum constant : constants) {
             try {
                 Field enumField = enumClass.getDeclaredField(constant.name());
-                if (!enumField.isAnnotationPresent(EnumValue.class))
-                    throw Exceptions.error("db enum must have @EnumValue, field={}, enum={}", Fields.path(field), Fields.path(enumField));
-                if (enumField.isAnnotationPresent(XmlEnumValue.class))
+                DBEnumValue enumValue = enumField.getDeclaredAnnotation(DBEnumValue.class);
+                if (enumValue == null) {
+                    throw Exceptions.error("db enum must have @DBEnumValue, field={}, enum={}", Fields.path(field), Fields.path(enumField));
+                }
+                boolean added = enumValues.add(enumValue.value());
+                if (!added) {
+                    throw Exceptions.error("db enum value must be unique, enum={}, value={}", enumClass.getCanonicalName() + "." + constant, enumValue.value());
+                }
+                if (enumField.isAnnotationPresent(XmlEnumValue.class)) {
                     throw Exceptions.error("db enum must not have jaxb annotation, please separate view and entity, field={}, enum={}", Fields.path(field), Fields.path(enumField));
+                }
             } catch (NoSuchFieldException e) {
                 throw new Error(e);
             }