+++ core-ng/src/main/java/core/framework/impl/mongo/LocalDateTimeCodec.java
@@ -1,10 +1,13 @@
 package core.framework.impl.mongo;
 
 import org.bson.BsonReader;
+import org.bson.BsonType;
 import org.bson.BsonWriter;
 import org.bson.codecs.Codec;
 import org.bson.codecs.DecoderContext;
 import org.bson.codecs.EncoderContext;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.time.Instant;
 import java.time.LocalDateTime;
@@ -14,21 +17,34 @@ import java.time.ZoneId;
  * @author neo
  */
 public class LocalDateTimeCodec implements Codec<LocalDateTime> {
+    private static final Logger LOGGER = LoggerFactory.getLogger(LocalDateTimeCodec.class);
+
     static void write(BsonWriter writer, LocalDateTime value) {
-        writer.writeDateTime(value.atZone(ZoneId.systemDefault()).toInstant().toEpochMilli());
+        if (value == null) writer.writeNull();
+        else writer.writeDateTime(value.atZone(ZoneId.systemDefault()).toInstant().toEpochMilli());
     }
 
     static LocalDateTime read(BsonReader reader) {
-        return LocalDateTime.ofInstant(Instant.ofEpochMilli(reader.readDateTime()), ZoneId.systemDefault());
+        BsonType currentType = reader.getCurrentBsonType();
+        if (currentType == BsonType.NULL) {
+            reader.readNull();
+            return null;
+        } else if (currentType == BsonType.DATE_TIME) {
+            return LocalDateTime.ofInstant(Instant.ofEpochMilli(reader.readDateTime()), ZoneId.systemDefault());
+        } else {
+            LOGGER.warn("field returned from mongo is ignored, field={}, type={}", reader.getCurrentName(), currentType);
+            reader.skipValue();
+            return null;
+        }
     }
 
     @Override
-    public void encode(BsonWriter writer, LocalDateTime value, EncoderContext encoderContext) {
+    public void encode(BsonWriter writer, LocalDateTime value, EncoderContext context) {
         write(writer, value);
     }
 
     @Override
-    public LocalDateTime decode(BsonReader reader, DecoderContext decoderContext) {
+    public LocalDateTime decode(BsonReader reader, DecoderContext context) {
         return read(reader);
     }