+++ core-ng/src/test/java/core/framework/impl/db/TestEnum.java
@@ -1,13 +1,13 @@
 package core.framework.impl.db;
 
-import core.framework.api.db.EnumValue;
+import core.framework.api.db.DBEnumValue;
 
 /**
  * @author neo
  */
 public enum TestEnum {
-    @EnumValue("DB_V1")
+    @DBEnumValue("DB_V1")
     V1,
-    @EnumValue("DB_V2")
+    @DBEnumValue("DB_V2")
     V2
 }